requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.2
certifi==2025.10.5
cffi==2.0.0
Django==5.2.6
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
minio==7.2.18
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.23
pycryptodome==3.23.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
